requirements.txt
streamlit==1.51.0
opencv-python==4.12.0.88
torch==2.5.1
torchvision==0.20.1
transformers>=4.37.0
pillow>=10.0.0
numpy>=1.26.0
safetensors==0.7.0